+++ .example.env
@@ -4,4 +4,5 @@ MONGO_INITDB_ROOT_PASSWORD=
 MONGO_INITDB_ROOT_USERNAME=
 NODE_ENV=
 PORT=
+RATE_LIMIT_TTL=
 REDIS_HOST=
