+++ docker-compose.yml
@@ -5,7 +5,7 @@ services:
       - .env
     build: .
     ports:
-      - "3000:3000"
+      - '3000:3000'
     volumes:
       - .:/src
       # prevents mapping host's node_modules to container
